+++ src/main/java/com/moviesApp/service/UserService.java
@@ -2,6 +2,8 @@ package com.moviesApp.service;
 
 import com.moviesApp.daoLayer.UserDAO;
 import com.moviesApp.entities.User;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 import java.sql.SQLException;
 import java.util.ArrayList;
@@ -12,47 +14,69 @@ import java.util.List;
  */
 public class UserService {
 
+    private static final Logger LOGGER = LogManager.getLogger();
+
     public User getUserByID(Long userID) {
+        if (userID <= 0) {
+            LOGGER.error("Failed to get user. User id must be > 0. User id: " + userID);
+            return null;
+        }
         UserDAO userDAO = new UserDAO();
         User user = new User();
         try {
             user = userDAO.get(userID);
         } catch (SQLException e) {
             e.printStackTrace();// TODO handle
+            LOGGER.error("SQLException: " + e.getMessage());
         }
         return user;
     }
 
     public Long createUser(String userName) {
+        if (userName == null || userName.trim().equals("")) {
+            LOGGER.error("Failed to create user. User name must not be null or empty. User name: " + userName);
+            return 0L;
+        }
         UserDAO userDAO = new UserDAO();
         Long userID = 0L;
         try {
             userID = userDAO.create(userName);
         } catch (SQLException e) {
             e.printStackTrace();// TODO handle
+            LOGGER.error("SQLException: " + e.getMessage());
         }
         return userID;
     }
 
     public void updateUser(Long userID, String userName) {
+        if (userID <= 0) {
+            LOGGER.error("Failed to get user. User id must be > 0. User id: " + userID);
+            return;
+        }
+        if (userName == null || userName.trim().equals("")) {
+            LOGGER.error("Failed to update user. User name must not be null or empty. User name: " + userName + ". User name set to origin.");
+            userName = getUserByID(userID).getName();
+        }
         UserDAO userDAO = new UserDAO();
         try {
-            if (userName == null || userName.trim().equals("")) {
-                System.out.println("Update failed due to wrong name");
-                return;
-            }
             userDAO.update(userID, userName);
         } catch (SQLException e) {
             e.printStackTrace();// TODO handle
+            LOGGER.error("SQLException: " + e.getMessage());
         }
     }
 
     public boolean deleteUser(Long userID) {
+        if (userID <= 0) {
+            LOGGER.error("Failed to get user. User id must be > 0. User id: " + userID);
+            return false;
+        }
         UserDAO userDAO = new UserDAO();
         try {
             return userDAO.delete(userID);
         } catch (SQLException e) {
             e.printStackTrace();// TODO handle
+            LOGGER.error("SQLException: " + e.getMessage());
         }
         return false;
     }
@@ -64,6 +88,7 @@ public class UserService {
             users = userDAO.getAll();
         } catch (SQLException e) {
             e.printStackTrace();// TODO handle
+            LOGGER.error("SQLException: " + e.getMessage());
         }
         return users;
     }
